feat(user): add user service and initial configurations
This commit adds a new service called user, which will contain all the functionalities related to a user, such as authentication, authorization, profile, etc.

This commit also adds the initial configurations of the user service, such as the models, serializers, views, and urls. The configurations are based on the requirements and specifications of the project.